Docker/Docker_Service_Face/ai-projects/camera-ai-service/pyproject.toml
[tool.poetry]
name = "camera-ai-service"
version = "0.1.0"
description = ""
authors = ["PhuongTanThanh <thanh.pt@oryza.vn>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pycurl = "7.45.3"
configparser = "7.0.0"
requests = "2.31.0"
paho-mqtt = "1.6.1"
pika = "1.3.2"
fastapi = "0.110.2"
minio = "7.2.5"
pydantic-settings = "2.2.1"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
python-multipart = "0.0.9"
pymongo = "4.7.2"
odmantic = "1.0.2"
pytest = "8.2.0"
httpx = "0.27.0"
numpy = "^1.26.4"
opencv-python = "^4.9.0.80"
annotated-types = "0.6.0"
python-version = "^0.0.2"
anyio = "4.3.0"
argon2-cffi-bindings = "21.2.0"
argon2-cffi = "23.1.0"
certifi = "2024.2.2"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
colorama = "0.4.6"
dnspython = "2.6.1"
h11 = "0.14.0"
httpcore = "1.0.5"
idna = "3.7"
iniconfig = "2.0.0"
motor = "3.4.0"
packaging = "24.0"
pluggy = "1.5.0"
pycparser = "2.22"
pycryptodome = "3.20.0"
pydantic-core = "2.18.2"
pydantic = "2.7.1"
python-dotenv = "1.0.1"
sniffio = "1.3.1"
starlette = "0.37.2"
typing-extensions = "4.11.0"
urllib3 = "2.2.1"
pillow = "^10.3.0"
requests-toolbelt = "^1.0.0"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
